Remove .envrc
Setting up Node is an environment-specific concern.  Case in point: I
use direnv to manage other environment settings and this direnv file
messes up my local environment. Adding `source_up` would fix this, but
adding `source_up` to catalog's `.envrc` broke direnv for Max. To me
this is a smell that this type of thing doesn't belong in repositories,
and in fact I get this functionality out of nodenv without committing
artifacts to repositories.