fix: always generate valid objectids
In https://github.com/goodeggs/fake-eggs/pull/50 we updated the default lower bound for
`fake.date()` from 10/04/2011 to -271821/04-20 (yes, you read that right).

However, we were using `fake.date()` with no arguments to generate the timestamp part of
MongoDB ObjectIDs in `fake.objectId()`, so _most_ of the time (when the generated date was
before the Unix epoch, 01/01/1970), we would end up with a negative timestamp, e.g.
`123456789`, so when we could convert this to a string and concatenate it with the other
parts, we would end up with something like `-539f74fa83b9a54d8d1ddbb`, which is in invalid
MongoDB ObjectID.