fix: move new jest setup file to test/ directory
Gross. TypeScript implicitly guesses the "root dir" of a project. See `rootDir` option:
https://www.typescriptlang.org/docs/handbook/compiler-options.html.

If all files share a root dir, e.g. `src/`, it will "guess" that the output directory
structure (say, if `outputDir` is `dist/`) should not include the root level, so `src/foo`
will become `dist/foo` rather than `dist/src/foo`.

So by including a new `.ts` file that was _not_ under the same "root" directory, the
default implicit TypeScript behavior was to assume `.` is now the root, which moved all
the output `.d.ts` type declaration files to the wrong location in the published module.
This results in us quietly publishing a completely untyped library due to a seemingly
innocuous change.

To prevent this from happening in the future, the next commit explicitly configures `rootDir`.